test case for spark engine
