+++ linkis-engineconn-plugins/spark/src/test/scala/org/apache/linkis/engineplugin/spark/executor/TestSparkSqlExecutor.scala
@@ -24,14 +24,15 @@ import org.apache.linkis.engineplugin.spark.entity.SparkEngineSession
 import org.apache.linkis.engineplugin.spark.factory.SparkEngineConnFactory
 import org.apache.spark.SparkConf
 import org.apache.spark.sql.SparkSession
-import org.junit.jupiter.api.{Assertions, Test}
+import org.junit.jupiter.api.{Assertions, BeforeAll, Test}
 
 import scala.collection.mutable
 
 class TestSparkSqlExecutor{
-  @Test
-  def testCreateContext: Unit = {
+  @BeforeAll
+  def initService: Unit = {
     System.setProperty("wds.linkis.server.version", "v1")
+    System.setProperty("wds.linkis.engineconn.plugin.default.class", "org.apache.linkis.engineplugin.spark.SparkEngineConnPlugin")
     val map = new mutable.HashMap[String, String]()
     map.put("spring.mvc.servlet.path", "/api/rest_j/v1")
     map.put("server.port", "26378")
@@ -39,6 +40,9 @@ class TestSparkSqlExecutor{
     map.put("eureka.client.register-with-eureka", "false")
     map.put("eureka.client.fetch-registry", "false")
     DataWorkCloudApplication.main(DWCArgumentsParser.formatSpringOptions(map.toMap))
+  }
+  @Test
+  def testCreateContext: Unit = {
     val engineFactory = new SparkEngineConnFactory
     val sparkConf = new SparkConf(true)
     System.setProperty("java.io.tmpdir", "./")
@@ -55,5 +59,27 @@ class TestSparkSqlExecutor{
     val code = "select * from temp"
     val response = sparkSqlExecutor.executeLine(engineExecutionContext, code)
     Assertions.assertNotNull(response)
+  }
+
+  @Test
+  def testShowDF: Unit = {
+    val engineFactory = new SparkEngineConnFactory
+    val sparkConf: SparkConf = new SparkConf(true)
+    System.setProperty("HADOOP_CONF_DIR", "./")
+    System.setProperty("wds.linkis.filesystem.hdfs.root.path", "./")
+    System.setProperty("java.io.tmpdir", "./")
+    val sparkSession = SparkSession.builder()
+      .master("local[*]")
+      .appName("testShowDF").getOrCreate()
+    val outputDir = engineFactory.createOutputDir(sparkConf)
+    val sparkEngineSession = SparkEngineSession(sparkSession.sparkContext, sparkSession.sqlContext, sparkSession, outputDir)
+    val sparkScalaExecutor = new SparkScalaExecutor(sparkEngineSession, 1L)
+    val engineExecutionContext = new EngineExecutionContext(sparkScalaExecutor, Utils.getJvmUser)
+    val dataFrame = sparkSession.createDataFrame(Seq(
+      ("ming", 20, 15552211521L),
+      ("hong", 19, 13287994007L),
+      ("zhi", 21, 15552211523L)
+    )).toDF("name", "age", "phone")
+    SQLSession.showDF(sparkSession.sparkContext, "test", dataFrame, "", 10, engineExecutionContext)
   }
 }
